Renaming ODE solvers and reorganising files
