+++ torchqdynamics/ode_adjoint_qsolver.py
@@ -7,173 +7,13 @@ import torch.nn as nn
 from torch import Tensor
 from torch.autograd.function import FunctionCtx
 
-from .adaptive import DormandPrince45
+from .ode_forward_qsolver import ODEForwardQSolver
 from .progress_bar import tqdm
-from .qsolver import QSolver
-from .solver_options import Dopri45, ODEAdaptiveStep, ODEFixedStep
+from .solver_options import ODEAdaptiveStep, ODEFixedStep
 from .solver_utils import add_tuples, none_to_zeros_like
 
 
-class ForwardQSolver(QSolver):
-    @abstractmethod
-    def forward(self, t: float, y: Tensor) -> Tensor:
-        """Iterate the quantum state forward.
-
-        Args:
-            t: Time.
-            y: Quantum state, of shape `(..., m, n)`.
-
-        Returns:
-            Tensor of shape `(..., m, n)`.
-        """
-        pass
-
-    def run(self):
-        """Integrate a quantum ODE starting from an initial state.
-
-        The ODE is solved from time `t=0.0` to `t=t_save[-1]`.
-        """
-        # check arguments
-        check_t_save(self.t_save)
-
-        # dispatch to appropriate odeint subroutine
-        if self.gradient_alg is None:
-            self._odeint_inplace()
-        elif self.gradient_alg == 'autograd':
-            self._odeint_main()
-
-    def _odeint_inplace(self):
-        """Integrate a quantum ODE with an in-place solver.
-
-        Simple solution for now so torch does not store gradients.
-        TODO Implement a genuine in-place solver.
-        """
-        with torch.no_grad():
-            self._odeint_main()
-
-    def _odeint_main(self):
-        """Dispatch the ODE integration to fixed or adaptive time step subroutines."""
-        if isinstance(self.options, ODEFixedStep):
-            self._fixed_odeint()
-        elif isinstance(self.options, ODEAdaptiveStep):
-            self._adaptive_odeint()
-
-    def _fixed_odeint(self):
-        """Integrate a quantum ODE with a fixed time step solver.
-
-        Note:
-            The solver times are defined using `torch.linspace` which ensures that the
-            overall solution is evolved from the user-defined time (up to an error of
-            `rtol=1e-5`). However, this may induce a small mismatch between the time
-            step inside `qsolver` and the time step inside the iteration loop. A small
-            error can thus buildup throughout the ODE integration. TODO Fix this.
-        """
-        # get time step
-        dt = self.options.dt
-
-        # assert that `t_save` values are multiples of `dt`
-        if not torch.allclose(torch.round(self.t_save / dt), self.t_save / dt):
-            raise ValueError(
-                'For fixed time step solvers, every value of `t_save` must be a'
-                ' multiple of the time step `dt`.'
-            )
-
-        # define time values
-        num_times = torch.round(self.t_save[-1] / dt).int() + 1
-        times = torch.linspace(0.0, self.t_save[-1], num_times)
-
-        # run the ode routine
-        y = self.y0
-        for t in tqdm(times[:-1], disable=not self.options.verbose):
-            # save solution
-            if t >= self.next_tsave():
-                self.save(y)
-
-            # iterate solution
-            y = self.forward(t, y)
-
-        # save final time step (`t` goes `0.0` to `t_save[-1]` excluded)
-        self.save(y)
-        self.save_final(y)
-
-    def _adaptive_odeint(self):
-        """Integrate a quantum ODE with an adaptive time step solver.
-
-        This function integrates an ODE of the form `dy / dt = f(t, y)` with
-        `y(0) = y0`, using a Runge-Kutta adaptive time step solver.
-
-        For details about the integration method, see Chapter II.4 of `Hairer et al.,
-        Solving Ordinary Differential Equations I (1993), Springer Series in
-        Computational Mathematics`.
-        """
-        # save initial solution
-        self.save(self.y0)
-
-        save_flag = False
-
-        # initialize the adaptive solver
-        args = (
-            self.forward,
-            self.options.factor,
-            self.options.min_factor,
-            self.options.max_factor,
-            self.options.atol,
-            self.options.rtol,
-        )
-        if isinstance(self.options, Dopri45):
-            solver = DormandPrince45(*args)
-
-        # initialize the ODE routine
-        t0 = 0.0
-        f0 = solver.f(t0, self.y0)
-        dt = solver.init_tstep(f0, self.y0, t0)
-
-        # initialize the progress bar
-        pbar = tqdm(total=self.t_save[-1].item(), disable=not self.options.verbose)
-
-        # run the ODE routine
-        t, y, ft = t0, self.y0, f0
-        step_counter, max_steps = 0, self.options.max_steps
-        next_tsave = self.next_tsave()
-        while t < self.t_save[-1] and step_counter < max_steps:
-            # if a time in `t_save` is reached, raise a flag and rescale dt accordingly
-            if t + dt >= next_tsave:
-                save_flag = True
-                dt_old = dt
-                dt = next_tsave - t
-
-            # perform a single solver step of size dt
-            ft_new, y_new, y_err = solver.step(ft, y, t, dt)
-
-            # compute estimated error of this step
-            error = solver.get_error(y_err, y, y_new)
-
-            # update results if step is accepted
-            if error <= 1:
-                t, y, ft = t + dt, y_new, ft_new
-
-                # update the progress bar
-                pbar.update(dt.item())
-
-                # save results if flag is raised
-                if save_flag:
-                    self.save(t, y)
-
-            # return to the original dt, lower the flag and get next save time
-            if save_flag:
-                dt = dt_old
-                save_flag = False
-                next_tsave = self.next_tsave()
-
-            # compute the next dt
-            dt = solver.update_tstep(dt, error)
-            step_counter += 1
-
-        # save last state if not already done
-        self.save_final(y)
-
-
-class AdjointQSolver(ForwardQSolver):
+class ODEAdjointQSolver(ODEForwardQSolver):
     GRADIENT_ALG = ['autograd', 'adjoint']
 
     def run(self):
@@ -221,7 +61,7 @@ class AdjointQSolver(ForwardQSolver):
 
 
 def _odeint_augmented_main(
-    qsolver: AdjointQSolver,
+    qsolver: ODEAdjointQSolver,
     y0: Tensor,
     a0: Tensor,
     g0: tuple[Tensor, ...],
@@ -241,7 +81,7 @@ def _adaptive_odeint_augmented(*_args, **_kwargs):
 
 
 def _fixed_odeint_augmented(
-    qsolver: AdjointQSolver,
+    qsolver: ODEAdjointQSolver,
     y0: Tensor,
     a0: Tensor,
     g0: tuple[Tensor, ...],
@@ -299,7 +139,7 @@ class ODEIntAdjoint(torch.autograd.Function):
     @staticmethod
     def forward(
         ctx: FunctionCtx,
-        qsolver: AdjointQSolver,
+        qsolver: ODEAdjointQSolver,
         y0: Tensor,
         t_save: Tensor,
         *parameters: tuple[nn.Parameter, ...],
@@ -370,16 +210,3 @@ class ODEIntAdjoint(torch.autograd.Function):
 
         # return the computed gradients w.r.t. each argument in `forward`
         return None, a, None, None, None, *g
-
-
-def check_t_save(t_save: Tensor):
-    """Check that `t_save` is valid (it must be a non-empty 1D tensor sorted in
-    strictly ascending order and containing only positive values)."""
-    if t_save.ndim != 1 or len(t_save) == 0:
-        raise ValueError('Argument `t_save` must be a non-empty 1D tensor.')
-    if not torch.all(torch.diff(t_save) > 0):
-        raise ValueError(
-            'Argument `t_save` must be sorted in strictly ascending order.'
-        )
-    if not torch.all(t_save >= 0):
-        raise ValueError('Argument `t_save` must contain positive values only.')
